Get unit tests going
