--- app/Format.py
@@ -1,6 +1,0 @@
-def clean(string):
-    return string.strip().lower()
-
-def dollar(amount):
-    return "$" + format(amount, '0,.2f')
-
