--- app/Utils.py
@@ -1,10 +1,0 @@
-BOOLEAN_LOOKUP = {
-    "y": True,
-    "n": False
-}
-
-ACCOUNT_LOOKUP = {
-    "c": "checking",
-    "s": "savings",
-    "m": "money market"
-}
