+++ src/Session.py
@@ -1,7 +1,8 @@
-from Utils import BOOLEAN_LOOKUP, ACCOUNT_LOOKUP
+"""
+"""
+
 from Customer import Customer
-from Format import clean, dollar
-# Should Utils and Format be combined?
+from Utils import BOOLEAN_LOOKUP, ACCOUNT_LOOKUP, clean, dollar
 
 class Session:
     def __init__(self, customer_list):
@@ -24,20 +25,30 @@ class Session:
         """
         print "Welcome to Austin Community Bank!"
         pin = str(raw_input("To get started, please enter your PIN: "))
+        attempts = 0
+        while len(pin) != 4:
+            warning = "Your PIN contains exactly 4 digits. Please try again: "
+            pin = str(raw_input(warning))
+            attempts += 1
+            if attempts >= 3:
+                print "Canceling transaction after three attempts."
+                break
+                return
         if pin == "xxxx":
             self.running = False
         else:
             self.current_customer = self.get_customer(pin)
 
     def get_customer(self, pin):
-        """Retrieve customer information by PIN."""
-        if len(pin) != 4:
-            raise ValueError("PIN should have exactly 4 digits.")
-        else:
-            return self.customers.get(pin,
-                "That is not a valid PIN. Please try again.")
+        """Retrieve customer information by PIN.
+        """
+        return self.customers.get(pin,
+            "That is not a valid PIN. Please try again.")
 
     def receipt(self):
+        """
+        """
+        self.current_customer.update_account_summary()
         if self.current_customer.num_transactions == 1:
             message = "You made {} transaction today. \n".format(
             self.current_customer.num_transactions)
@@ -48,8 +59,8 @@ class Session:
         return message
 
     def login(self):
-        # In real life, this would be a credit card number, and we would
-        # confirm via PIN
+        """
+        """
         print "Hello, {}!".format(self.current_customer.name)
         print self.current_customer.display_account_summary()
 
@@ -77,7 +88,8 @@ class Session:
             print "Not a valid response. Canceling transaction."
             return
     def serve(self):
-        # Count number of transactions. (Do I need this?)
+        """
+        """
         decision = self.proceed(first_time=True)
         while decision is True:
             account_choice = clean(raw_input(self.current_customer.display_account_choices()))
@@ -88,4 +100,6 @@ class Session:
         print self.receipt()
 
     def logout(self):
+        """
+        """
         print "Thank you for banking with Austin Community Bank. Good bye!"
